backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
ultralytics
pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
google-generativeai==0.8.3
python-dotenv==1.0.1
shapely
